refactor PlaylistHeader, Tracklist, and ActiveTrackSeeker into a MainView component; create Sidebar component
