+++ src/App/Playlist/Main/ActiveTrackSeeker/ActiveTrackSeeker.scss
