+++ src/App/Playlist/Main/Main.scss
@@ -1,0 +1,2 @@
+.playlist-main-container {
+}
