+++ src/App/Playlist/Main/PlaylistHeader/PlaylistHeader.scss
