+++ src/App/Playlist/Main/Tracklist/TrackContainer/AlbumCover/AlbumCover.scss
