+++ src/App/Playlist/Main/Tracklist/TrackContainer/MainInfo/MainInfo.scss
