+++ src/App/Playlist/Main/Tracklist/TrackContainer/ProgressBar/ProgressBar.scss
