+++ src/App/Playlist/Playlist.scss
@@ -1,0 +1,5 @@
+.playlist-container {
+  .loaded-playlist-container {
+    display: flex;
+  }
+}
