+++ src/App/Playlist/Sidebar/Sidebar.scss
@@ -1,0 +1,2 @@
+.sidebar {
+}
