+++ src/App/Playlist/Sidebar/index.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import './Sidebar.scss'
+
+const Sidebar = () => <div className='sidebar'>SIDEBAR</div>
+
+export default Sidebar
